app/src/test/java/com/dicoding/bengkelsaku/main/MainViewModelTest.kt
package com.dicoding.bengkelsaku.viewApp.main

import androidx.arch.core.executor.testing.InstantTaskExecutorRule
import androidx.lifecycle.LiveData
import androidx.lifecycle.MutableLiveData
import androidx.paging.AsyncPagingDataDiffer
import androidx.paging.PagingData
import androidx.paging.PagingSource
import androidx.paging.PagingState
import androidx.recyclerview.widget.ListUpdateCallback
import com.dicoding.bengkelsaku.DataDummy
import com.dicoding.bengkelsaku.MainDispatcherRule
import com.dicoding.bengkelsaku.data.UserRepository
//import com.dicoding.bengkelsaku.data.adapter.StoryAdapter
import com.dicoding.bengkelsaku.getOrAwaitValue
//import com.dicoding.bengkelsaku.response.Story
import kotlinx.coroutines.Dispatchers
import kotlinx.coroutines.ExperimentalCoroutinesApi
import kotlinx.coroutines.test.runTest
import org.junit.Assert
import org.junit.Assert.*
import org.junit.Rule
import org.junit.Test
import org.junit.runner.RunWith
import org.mockito.Mock
import org.mockito.Mockito
import org.mockito.junit.MockitoJUnitRunner

@ExperimentalCoroutinesApi
@RunWith(MockitoJUnitRunner::class)
class MainViewModelTest{
    @get:Rule
    val instantExecutorRule = InstantTaskExecutorRule()
    @get:Rule
    val mainDispatcherRules = MainDispatcherRule()
    @Mock
    private lateinit var quoteRepository: UserRepository
    private val Authorization =  "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJ1c2VyLUpJdUhrTGlucDY0WlJmV2UiLCJpYXQiOjE2OTkxNzM4NjF9.1vENjuQ2DP8yS-16fZBb4ReHMXwSz0gEolap_eHFeBQ"

//    @Test
//    fun `when Get Story Should Not Null and Return Data`() = runTest {
//        val dummyQuote = DataDummy.generateDummyQuoteResponse()
//        val data: PagingData<Story> = QuotePagingSource.snapshot(dummyQuote)
//        val expectedQuote = MutableLiveData<PagingData<Story>>()
//        expectedQuote.value = data
//        Mockito.`when`(quoteRepository.getPaging(Authorization)).thenReturn(expectedQuote)
//
//        val mainViewModel = MainViewModel(quoteRepository)
//        val actualQuote: PagingData<Story> = mainViewModel.story(Authorization).getOrAwaitValue()
//
//        val differ = AsyncPagingDataDiffer(
//            diffCallback = StoryAdapter.DIFF_CALLBACK,
//            updateCallback = noopListUpdateCallback,
//            workerDispatcher = Dispatchers.Main,
//        )
//        differ.submitData(actualQuote)
//
//        Assert.assertNotNull(differ.snapshot())
//        Assert.assertEquals(dummyQuote.size, differ.snapshot().size)
//        Assert.assertEquals(dummyQuote[0], differ.snapshot()[0])
//
//    }
//    @Test
//    fun `when Get Story Empty Should Return No Data`() = runTest {
//        val data: PagingData<Story> = PagingData.from(emptyList())
//        val expectedStory = MutableLiveData<PagingData<Story>>()
//        expectedStory.value = data
//        Mockito.`when`(quoteRepository.getPaging(Authorization)).thenReturn(expectedStory)
//        val mainViewModel = MainViewModel(quoteRepository)
//        val actualStory: PagingData<Story> = mainViewModel.story(Authorization).getOrAwaitValue()
//        val differ = AsyncPagingDataDiffer(
//            diffCallback = StoryAdapter.DIFF_CALLBACK,
//            updateCallback = noopListUpdateCallback,
//            workerDispatcher = Dispatchers.Main,
//        )
//        differ.submitData(actualStory)
//        Assert.assertEquals(0, differ.snapshot().size)
//    }
}

//class QuotePagingSource : PagingSource<Int, LiveData<List<Story>>>() {
//    companion object {
//        fun snapshot(items: List<Story>): PagingData<Story> {
//            return PagingData.from(items)
//        }
//    }
//    override fun getRefreshKey(state: PagingState<Int, LiveData<List<Story>>>): Int {
//        return 0
//    }
//    override suspend fun load(params: LoadParams<Int>): LoadResult<Int, LiveData<List<Story>>> {
//        return LoadResult.Page(emptyList(), 0, 1)
//    }
//}
val noopListUpdateCallback = object : ListUpdateCallback {
    override fun onInserted(position: Int, count: Int) {}
    override fun onRemoved(position: Int, count: Int) {}
    override fun onMoved(fromPosition: Int, toPosition: Int) {}
    override fun onChanged(position: Int, count: Int, payload: Any?) {}
}
